Addressing Duo's comments
